#208 let the db engine resolve queue operations
