+++ src/cmd/socket/nodepb/messages_pb.js
@@ -7158,7 +7158,6 @@ proto.pb.Message = {
   MESSAGE_QUEUE_LIST: 300,
   MESSAGE_QUEUE_ADD: 301,
   MESSAGE_QUEUE_REMOVE: 302,
-  MESSAGE_QUEUE_SAVE: 303,
   MESSAGE_DIRECTORY_READ: 400,
   MESSAGE_MICRO_SERVICE_ARTWORK: 420,
   MESSAGE_MICRO_SERVICE_AUDIOSCAN: 421,
